+++ internal/commands/networkpeering/delete.go
@@ -5,8 +5,10 @@ import (
 
 	"github.com/UpCloudLtd/upcloud-cli/v3/internal/commands"
 	"github.com/UpCloudLtd/upcloud-cli/v3/internal/completion"
+	"github.com/UpCloudLtd/upcloud-cli/v3/internal/config"
 	"github.com/UpCloudLtd/upcloud-cli/v3/internal/output"
 	"github.com/UpCloudLtd/upcloud-cli/v3/internal/resolver"
+	"github.com/spf13/pflag"
 
 	"github.com/UpCloudLtd/upcloud-go-api/v8/upcloud/request"
 )
@@ -27,27 +29,35 @@ type deleteCommand struct {
 	*commands.BaseCommand
 	resolver.CachingNetworkPeering
 	completion.NetworkPeering
+
+	disable config.OptionalBoolean
 }
 
 // InitCommand implements Command.InitCommand
 func (c *deleteCommand) InitCommand() {
+	flags := &pflag.FlagSet{}
+	config.AddToggleFlag(flags, &c.disable, "disable", false, "Disable the network peering before deleting it. This is equivalent to running 'upctl network-peering disable --wait` before 'upctl network-peering delete'.")
+	c.AddFlags(flags)
+
 	// Deprecating networkpeering in favour of network-peering
 	// TODO: Remove this in the future
 	commands.SetSubcommandDeprecationHelp(c, []string{"networkpeering"})
 }
 
-// Execute implements commands.MultipleArgumentCommand
-func (c *deleteCommand) Execute(exec commands.Executor, arg string) (output.Output, error) {
-	// Deprecating networkpeering in favour of network-peering
-	// TODO: Remove this in the future
-	commands.SetSubcommandExecutionDeprecationMessage(c, []string{"networkpeering"}, "network-peering")
+func Delete(exec commands.Executor, uuid string, disable bool) (output.Output, error) {
+	if disable {
+		_, err := disablePeering(exec, uuid, true)
+		if err != nil {
+			return nil, err
+		}
+	}
 
 	svc := exec.All()
-	msg := fmt.Sprintf("Deleting network peering %v", arg)
+	msg := fmt.Sprintf("Deleting network peering %s", uuid)
 	exec.PushProgressStarted(msg)
 
 	err := svc.DeleteNetworkPeering(exec.Context(), &request.DeleteNetworkPeeringRequest{
-		UUID: arg,
+		UUID: uuid,
 	})
 	if err != nil {
 		return commands.HandleError(exec, msg, err)
@@ -56,4 +66,13 @@ func (c *deleteCommand) Execute(exec commands.Executor, arg string) (output.Outp
 	exec.PushProgressSuccess(msg)
 
 	return output.None{}, nil
+}
+
+// Execute implements commands.MultipleArgumentCommand
+func (c *deleteCommand) Execute(exec commands.Executor, arg string) (output.Output, error) {
+	// Deprecating networkpeering in favour of network-peering
+	// TODO: Remove this in the future
+	commands.SetSubcommandExecutionDeprecationMessage(c, []string{"networkpeering"}, "network-peering")
+
+	return Delete(exec, arg, c.disable.Value())
 }