+++ CHANGELOG.md
@@ -7,9 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-### Fixed
+### Added
+
+- Experimental support for listing and deleting all resources with `all list` and `all purge` commands. This initial version supports networks, network-peerings, routers, and object-storages.
+- Add `--wait` flag to `object-storage delete` command.
 
 ## [3.16.1] - 2025-03-07
+
+### Fixed
 
 - Remove client side default value for kubernetes cluster plan and use default from API instead if no plan is defined.